requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.9.2
pydantic-settings==2.4.0
sqlalchemy==2.0.34
psycopg[binary]==3.2.3
pgvector==0.2.5
alembic==1.13.2
python-frontmatter==1.0.0
markdown-it-py==3.0.0
pdfminer.six==20231228
mail-parser==3.15.0
python-dateutil==2.9.0.post0
httpx==0.27.0
orjson==3.10.7
tenacity==9.0.0
apscheduler==3.10.4
python-dotenv==1.0.1
tantivy==0.25.0
jinja2==3.1.4
jsonschema==4.23.0
uvloop==0.19.0; sys_platform != "win32"
ruff==0.5.7
pytest==8.3.2
pyyaml==6.0.2
